move sql/dynamo tables into single file to make PR review easier
